pyproject.toml
[project]
name = "gcp-document-ai-data-parsing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-cloud-documentai>=3.7.0",
    "google-cloud-storage>=3.6.0",
    "numpy>=2.3.5",
    "opencv-python>=4.11.0.86",
    "pdf2image>=1.17.0",
    "pillow>=12.0.0",
    "tqdm>=4.67.1",
]
